Deevio_ML/requirements.txt
http://download.pytorch.org/whl/cpu/torch-0.3.1-cp35-cp35m-linux_x86_64.whl
numpy
scipy
opencv-python
scikit-learn
joblib
Pillow
h5py
ipython
wget
requests
statsd
scikit-image
tqdm
matplotlib
flask
torchvision
urllib3
imutils